feat: Add support for different application environments
This commit updates the code to support different application environments. It adds a new `APPLICATION_ENV` environment variable and sets its value to "production" in the build workflow. The `google_auth.go` file has been modified to read the `APPLICATION_ENV` variable from the configuration file. In the `google_client.go` file, the logic for creating the `ImageAnnotatorClient` has been updated to check the `APPLICATION_ENV` and use the appropriate credential file based on the environment.